Aula 2

+++ app1/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class User(models.Model):
     name = models.CharField("name", max_length=30)
-    phone = models.IntegerField("phone")
+    phone = models.BigIntegerField("phone")
     email = models.EmailField("email", max_length=30)
 
     def __str__(self):
